The GuiCluster & GuiWindow extend GuiMenuComponent, that combines common code.
